Config: Port GPUDeterminismMode setting to new config system.
